More warning fixes
Former-commit-id: d19df9cadfec8af9fca3ce046f46a825d60c0e3b

+++ Sample/source/Model.cpp
@@ -452,10 +452,10 @@ void Model::Draw(MaterialProgram *program, const Node& _skeleton)
         }
         
         // Pass the bone count through (well, anything more than zero means skinned)
-        glUniform1i(program->UniformSkinned(), (*renderable)->bones.size());
+        glUniform1i(program->UniformSkinned(), (GLint)(*renderable)->bones.size());
         
         // Pass the texture count through (anything more than zero means skinned)
-        glUniform1i(program->UniformTextured(), (*renderable)->material->Textures().size());
+        glUniform1i(program->UniformTextured(), (GLint)(*renderable)->material->Textures().size());
         
         
         // Set the reflectivity (should pull this from the material right??)
